+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## 0.3.0
 
 * Handle the i18next events
-* Support i18next methods addResourceBundle, addResourceBundle, and removeResourceBundle
+* Support i18next methods addResourceBundle, addResourceBundle, getDataByLanguage, and removeResourceBundle
 * Support store events
 * Method chaining from method use